% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3utilities.R
\name{add.dim}
\alias{add.dim}
\title{Adds a new dimension to an MD3 array}
\usage{
add.dim(x, .dimname = "XY", .dimcodes = NULL, .fillall = FALSE, ...)
}
\arguments{
\item{x}{an md3 object}

\item{.dimname}{the name of the new dimension. Has to be a single character element}

\item{.dimcodes}{an optional character vector with the code names of the new elements in that dimension}

\item{.fillall}{if FALSE and length(.dimcodes)>1 then the first element will contain the existing data while the others will be NA
if TRUE and length(.dimcodes)>1 then each new element will be filled with the same data as the first one}

\item{\dots}{unused}
}
\value{
an md3
}
\description{
Adds new dimension to an MD3 object, optionally to be filled with a number of new elements
}
\examples{
data("euhpq")
p1=add.dim(euhpq,'Residence',c('Primary','Secondary'))
print(p1[,1,1,4,])
p2=add.dim(euhpq,'Residence',c('Primary','Secondary'),.fillall=TRUE)
print(p2[,1,1,4,])

}
