% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3utilities.R
\name{aggregate.md3}
\alias{aggregate.md3}
\title{Agreggate an MD3 over time and/or multiple other dimensions}
\usage{
\method{aggregate}{md3}(
  x,
  frq_grp,
  along = "TIME",
  FUN = c(sum, mean, end, start),
  na.rm = TRUE,
  ...,
  complete.periods = !na.rm
)
}
\arguments{
\item{x}{an md3 object,}

\item{frq_grp}{what to aggregate to: A single character frequency code like \code{"A"},  \code{"Q"} or  \code{"B"} when aggregating along time, or a user-defined aggregate list (see details)}

\item{along}{characer vector: the dimensions along which to aggregate. Default is \code{TIME}}

\item{FUN}{the function to use for aggregation. Default is \code{sum}, but \code{mean},  \code{end} and  \code{start} are also popular. Can be any function (passed along as a function or the string name of that function)}

\item{na.rm}{logical, default \code{TRUE}. Should missing values be removed?}

\item{\dots}{other arguments to FUN}

\item{complete.periods}{logical. Applies only in case of time aggregation and \code{na.rm=FALSE}. if e.g. \code{FALSE} and \code{FUN=end} then this takes the last value of the last available subperiod in case x ends within a period  (see examples).}
}
\value{
an md3 object. Note that any flags or observation metadata from the original MD3 will be dropped and not contained in the resulting md3.
}
\description{
Can sum, average, etc. to lower frequencies, or along entire dimensions, or along user-defined aggregates (such as country groups)
}
\examples{
data("oecdgdp_aq")

yy=oecdgdp_aq['Q','AU:BE',,'2011q2:2015q2']
yy[AU..y2011q2+y2011q3]=NA #make some artifical NAs for demo purposes
print(yy[,'GDP_USD',])
aggregate(yy,'A')[,'GDP_USD',] # sum quarterly GDP figures to annual ones
aggregate(yy,'A',FUN=sum, na.rm=FALSE)[,'GDP_USD',] # sum quarterly GDP figures to annual ones.
#Note the NAs for 2011 and 2015 because 2011q1 and 2015q3:2015q4 are missing completely

Alternatively sum those years where not all periods are available by setting complete.periods=FALSE
aggregate(yy,'A',FUN=sum, na.rm=FALSE, complete.periods=FALSE)[,'GDP_USD',]


aggregate(yy,'A',FUN=end)[,'GDP_USD',] use the last period only
aggregate(yy,'A',FUN=end, na.rm=FALSE)[,'GDP_USD',] use the last period only, but be careful about NAs


}
