% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3timo64.R
\name{as.timo}
\alias{as.timo}
\alias{timo}
\title{A time period class for mixed frequencies}
\usage{
as.timo(x, frq = NULL, ...)

timo(..., frq = NULL)
}
\arguments{
\item{x}{vector to be converted to timd vector. Can be character, date, PSOIXct, etc.}

\item{frq}{frequency code like "A" or "M" or "B"}

\item{...}{other arguments passed on to subsequent methods}
}
\description{
A time period class for mixed frequencies
}
\details{
The timo object is a vector denoting mixed frequency periods like '2013', '2013m06', or '2013-06-24'.

The following denotes the existing frequency codes and their notation
\itemize{
\item \code{A} annual, notation \code{'2011'}
\item \code{S} semi-annual, notation \code{'2011s2'}, second half of 2011
\item \code{Q} quarterly, notation \code{'2011q3'}, third quarter of 2011
\item \code{M} monthly, notation \code{'2011m10'}, October 2011
\item \code{W} weekly, notation \code{'2011w39'}, 39th week of 2011
\item \code{B} business-day, notation \code{'2011-10-05'}, 5 October 2011
\item \code{D} daily, notation \code{'2011-10-02'}, 2 October 2011
\item \code{N} minutely, notation \code{'2011-10-02t13:24'}, 2 October 2011, 13:24h (default GMT)
}
}
\section{Functions}{
\itemize{
\item \code{timo()}: convert to timo object of time periods

}}
\examples{
timo(c('2011q1','2012q1'))
as.timo(c('2011q1','2012q1'))
as.timo('2011q1',frq='M')
as.timo('2011q1') + 0:8

as.timo(c('2011q1', '2011')) + 1
as.timo('2011/9/23')
timo('2011/9/23','Q')

timo(y2019q4) -'2q' #substract 2 quarters
timo(y2019q4) -'2a' #substract 2 years
timo("2019q4") +'12m' #add 12 months

seq(as.timo('2011m03'),'2012m01')

sort(as.timo(c("2012m01","2011","2011q2","2011s2","2011m03","2011w51",
   "2011-08-23","2012", '2011-08-23 11:11')))

sort(as.timo('2011q1') + 0:8, decreasing=TRUE)

}
