% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3utilities.R
\name{cfc}
\alias{cfc}
\alias{ctc}
\title{Utility functions - copy to-from clipbaord}
\usage{
cfc(
  header = TRUE,
  row.names = NULL,
  na.strings = c("", "#N/A", "#NUM!", "#REF!", "#VALUE!", "#DIV/0!", "#NAME?"),
  sep = "\\t",
  ...
)

ctc(x, ..., row.names = TRUE, col.names = TRUE, na = "")
}
\arguments{
\item{row.names}{whether to copy rownames}

\item{na.strings}{a characer vector that identifies NAs. Default is to recognize Excel NAs}

\item{sep}{which separator to use when copying form clipboard default is (Windows) tab}

\item{...}{any other arguments to \code{\link[utils]{write.table}} , respectively \code{\link[utils]{read.table}}}

\item{x}{any object containing data, such a data.frame, data.table, array, matrix, or  md3 object,}

\item{col.names}{whether to copy colnames (for ctc)}

\item{na}{how to copy NA. E.g. \code{=na()} for copying to Excel}
}
\description{
Can decompose sums, averages, etc. to higher frequencies, or to constituent parts (e.g. countries in country groups)
}
\examples{

ctc(matrix(1:4,2),row.names = FALSE)
cfc()

}
