% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3object.R
\name{dimcodes}
\alias{dimcodes}
\alias{dimcodes<-}
\alias{dimcodes.default}
\alias{dimcodes<-.default}
\alias{dimcodes.md3}
\alias{dimcodes<-.md3}
\title{Set/get dimension and element names, labels and attributes}
\usage{
dimcodes(x, ...)

dimcodes(x, ...) <- value

\method{dimcodes}{default}(x, ...)

\method{dimcodes}{default}(x) <- value

\method{dimcodes}{md3}(x, ...)

\method{dimcodes}{md3}(x) <- value
}
\arguments{
\item{x}{an md0 object or any base object}

\item{...}{no effect}

\item{value}{sets dimcodes equal to value}
}
\value{
list of dimcodes (for getting) resp. md0/generic object (for setting)
}
\description{
Set/get dimension and element names, labels and attributes
}
\details{
\code{dimcodes} is an attribute of an md0 that is a list with one data.frame per dimnesion name.
  Each data.frame contains at least one column that holds the "code"/element name for that identifier (e.g. country codes).
  The other columns hold attributes for those codes, most notable labels/descriptions.

  Note that R will automatically generate the column label:en, respectively label:YOURLANGUAGE to hold such labels
  Note that all such labels need to follow the structure label:LANG

  You can add arbitrary columns and with attributes for each code.
  The following column names, however, have special meaning: 1) \code{label:LANG}, 2) \code{parent}
}
\section{Methods (by class)}{
\itemize{
\item \code{dimcodes(default)}: Returns \code{dimnames} for generic objects

\item \code{dimcodes(md3)}: Returns dimcodes for md0 objects

}}
\section{Functions}{
\itemize{
\item \code{dimcodes(default) <- value}: Sets \code{dimnames} for generic objects

\item \code{dimcodes(md3) <- value}: Sets dimcodes for md0 objects

}}
\examples{
data(euca)
names(dimcodes(euca)) # dimension names
dimcodes(euca) # element names
dimcodes(euca)[[1]][,"label:fr"] = c("Millions euro","\% PIB") #add labels in French language
dimcodes(euca)[[2]][1:3,2] = paste(dimcodes(euca)[[2]][1:3,2], "cool") #adjust labels
dimcodes(euca)[["unit"]]["PC_GDP",1] = "PC" #rename element
names(dimcodes(euca))[[2]] = "country" #rename dimension
}
\seealso{
\code{\link{dimnames}}
}
