% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3object.R
\name{.fixfactors}
\alias{.fixfactors}
\title{Converts factors to strings in a data.frame or list}
\usage{
.fixfactors(data)
}
\arguments{
\item{data}{a data .frame or other list}
}
\value{
a data.frame/list
}
\description{
Converts factors to strings in a data.frame or list
}
\examples{
data("ToothGrowth")
ToothGrowth[["supp"]]
tg1=.fixfactors(ToothGrowth)
tg1[,"supp"]
}
\seealso{
\code{\link{as.character}}
}
