% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3object.R
\name{.mdcodes2rest}
\alias{.mdcodes2rest}
\title{Concatenates columns in a matrix into a SXDM Restful code equivalent}
\usage{
.mdcodes2rest(x)
}
\arguments{
\item{x}{a data.frame or matrix}
}
\value{
a character vector
}
\description{
Concatenates columns in a matrix into a SXDM Restful code equivalent
}
\examples{
dd=cbind(LETTERS,letters,rev(letters))
.mdcodes2rest(dd)
}
