% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3object.R
\name{drop.md3}
\alias{drop.md3}
\title{Dropping dimensions}
\usage{
\method{drop}{md3}(x, ...)
}
\arguments{
\item{x}{an md3 object with n dimensions}
}
\value{
an md3 object with n dimensions or less
}
\description{
Drops any dimensions that contain only a single element
}
\examples{
#data(euhpq)
ww=euhpq[TOTAL.I15_Q..,drop=FALSE]
dim(ww)# dimensions
w2=drop(ww)
dim(w2)
}
\seealso{
\code{\link{aperm.md3}}, \code{\link{add.dim}}
}
