% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3utilities.R
\name{fill}
\alias{fill}
\title{Fill out an MD3 or array with a number or a certain subset}
\usage{
fill(bigmd3, ...)
}
\arguments{
\item{bigmd3}{an md3 object, array, or anything that may be converted to an md3}

\item{\dots}{a numeric, array, or md3 with which to fill out all values of bigmd3. Alternatively, the indexing reference to a subset of the md3 (see \link{indexMD3})}
}
\value{
an md3 object (or an array if bigmd3 is an array)
}
\description{
This function is useful when one wants to use the same sequence of numbers along a certain dimension, e.g. when computing an index
}
\examples{
data("euhpq")

mm=euhpq[1,1,1:4,as.character(2013:2016)]

# these four are equivalent:
fill(mm,mm[,'2014q4'])
fill(mm,'.2014q4')
fill(mm,'BE+BG.2014q4')
fill(mm,c('BE','BG'),'2014q4')

mm/fill(mm,rowMeans(mm[,'2014']))*100 # e.g. re-base the index with the av of 2014 being 100

fill(mm,100)
fill(mm,c('AT'=1,'BG'=2))

fill(mm,c('2014q1'=1,'2014q2'=2))


fill(mm,1:2)



}
\seealso{
\code{\link{aggregate.md3}}, \link{indexMD3}, \code{\link{imputena}}
}
