% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3utilities.R
\name{frequency.md3}
\alias{frequency.md3}
\alias{frequency<-.md3}
\title{Getting, setting or converting the frequency of an MD3 object}
\usage{
\method{frequency}{md3}(x, ...)

\method{frequency}{md3}(x, refersto = "end") <- value
}
\arguments{
\item{x}{an \code{md3} object}

\item{value}{the frequency to which the object is to be changed to, expressed as a single character such as \code{'Q'}}

\item{refersto}{an optional character \code{'start'}, \code{'middle'}, or \code{'end'}. If \code{frq} is higher frequency than the original x, then this chooses the first a last eligible subperiod.}
}
\value{
A character vector with the unique frequency codes (like 'D','B','A') that \code{x} contains. Empty character if \code{x} contains not TIME dimension
}
\description{
Getting, setting or converting the frequency of an MD3 object
}
\details{
The following denotes the existing frequency codes and their notation
\itemize{
\item \code{A} annual, notation \code{'2011'}
\item \code{S} semi-annual, notation \code{'2011s2'}, second half of 2011
\item \code{Q} quarterly, notation \code{'2011q3'}, third quarter of 2011
\item \code{M} monthly, notation \code{'2011m10'}, October 2011
\item \code{W} weekly, notation \code{'2011w39'}, 39th week of 2011
\item \code{D} daily, notation \code{'2011-10-02'}, 2 October 2011
\item \code{N} minutely, notation \code{'2011-10-02t13:24'}, 2 October 2011, 13:24h (default GMT)
}
}
\examples{
frequency(eupop)

}
\seealso{
\code{\link{frequency.timo}}, \code{\link{timo}}
}
