% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3timo64.R
\name{frequency.timo}
\alias{frequency.timo}
\alias{frequency<-.timo}
\title{Getting, setting or converting the frequency of a timo period object}
\usage{
\method{frequency}{timo}(x, frq = NULL, refersto = "start", ...)

\method{frequency}{timo}(x, refersto = c("start", "end")) <- value
}
\arguments{
\item{x}{a \code{\link{timo}}  period vector}

\item{frq}{a frequency character like '\code{A}', 'M' or 'D' to convert x to}

\item{refersto}{an optional character vector. If frq is higher frequency than the original x, then this chooses the first a last eleigible subperiod.}
}
\value{
a timo obejct if frq is specifed. A character vector with frequency codes (like 'D','B','A') otherwise
}
\description{
Getting, setting or converting the frequency of a timo period object
}
\details{
The following denotes the existing frequency codes and their notation
\itemize{
\item \code{A} annual, notation \code{'2011'}
\item \code{S} semi-annual, notation \code{'2011s2'}, second half of 2011
\item \code{Q} quarterly, notation \code{'2011q3'}, third quarter of 2011
\item \code{M} monthly, notation \code{'2011m10'}, October 2011
\item \code{W} weekly, notation \code{'2011w39'}, 39th week of 2011
\item \code{D} daily, notation \code{'2011-10-02'}, 2 October 2011
\item \code{N} minutely, notation \code{'2011-10-02t13:24'}, 2 October 2011, 13:24h (default GMT)
}
}
\examples{
frequency(timo(c('2011q1','2023m11','2014-12-09')))
aa <- timo(y2022q3,y2022q4)
frequency(aa)
frequency(aa) <- 'm'
aa
frequency(aa,'D')
frequency(aa,frq='D',refersto='end')
frequency(seq(timo(2013),2014,by='Q'),frq='D',refersto=c('start','end'))
}
