% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3utilities.R
\name{\%&\%}
\alias{\%&\%}
\title{Utility functions - paste operator}
\usage{
... \%&\% collapse
}
\arguments{
\item{...}{things to paste}
}
\description{
Pastes two strings without having to type paste
}
\examples{

"A" \%&\% "B"
LETTERS \%&\% 2

}
