% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3utilities.R
\name{imputena}
\alias{imputena}
\title{Impute NAs in MD3 objects with proxy data or interpolation}
\usage{
imputena(
  x,
  proxy = NULL,
  method = c("dlog", "diff"),
  maxgap = 6,
  direction = c("both", "forward", "backward")
)
}
\arguments{
\item{x}{an md3 object}

\item{proxy}{either a vector (applied to each time series), or an array or an md3 with the same dimensions or dimension names as x. If proxy left to NULL then this function will merely interpolate any gaps it finds}

\item{method}{'dlog' denotes geometric interpolation, 'diff' denotes linear interpolation}

\item{maxgap}{an integer denoting the maximum number of consecutive NAs to be filled}

\item{direction}{only necessary where proxy is not NULL. See details below}
}
\value{
an md3 object
}
\description{
For each time series in the MD3 array, this function fills intermittent NAs using the delta of a proxy, or (geometric) interpolation
}
\details{
This refers to interpolation cases where an annual time series e.g. has values for 2000 and 2005, but not for 2001:2004

If maxgap =3 or more, then interpol will fill the NAs of 2001:2004

Note that interpolna does not work for mixed frequencies

As regards imputation by proxy, the function will impute NAs in time series that have a least one non-missing value with growth rates
or deltas of the proxy. direction=='forward' means it does this for any NAs after non-missing observations, direction==' backward' for
any NAs before. Direction 'both' is a weighted avearge of both methods wherever they overlap
}
\examples{
data("euhpq")
\dontrun{w1= euhpq[TOTAL.I15_Q.AT:CY.y2005:y2008] #make small md3}
w1["BE.2006q4:2007q1"]=NA
w1["CY.2006:2007"]=NA
w1['AT.y2006q1']=100
print(w1)
#just interpolate
imputena(w1) #note the difference 'maxgap'  makes for CY here
imputena(w1,maxgap=10) #note the difference 'maxgap'  makes for CY here



using a lower-dimensional proxy
\dontrun{imputena(w1, euhpq['TOTAL.I10_Q.AT:CY.y2005:y2008'])}

#with trend
imputena(w1, 1:16)

}
