% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3object.R
\name{is.na.md3}
\alias{is.na.md3}
\title{Apply is.na to md3 objects}
\usage{
\method{is.na}{md3}(x, asarray = TRUE)
}
\arguments{
\item{x}{an md3 object or any base object}

\item{asarray}{if this is TRUE, the function behavres like is.na on an array and returns an array. if \code{FALSE} then the funciton returns a vecotr, which is faster}
}
\value{
logical vector or array of the same dimensions as x
}
\description{
Apply is.na to md3 objects
}
\details{
The sparse formulation of md3 obejcts means that, under the hood, NAs are thos edimension names that are not present in the underlying data.table.
For this reason, md3 require an is.na function that is specfic to them
}
\examples{
aa=eupop['NL:PT..TOTAL.1984']

is.na(aa)
is.na(aa,FALSE)

aa[is.na(aa)]<-0
#aa[onlyna=TRUE]<-0 # would have been faster
}
\seealso{
\code{\link{dimnames}}
}
