% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3utilities.R
\name{lag.md3}
\alias{lag.md3}
\alias{diff.md3}
\alias{growth}
\title{lag MD3 objects}
\usage{
\method{lag}{md3}(x, k = 1, whichdim = length(dim(x)), constrange = TRUE, na.pad = TRUE, ...)

\method{diff}{md3}(
  x,
  lag = 1L,
  whichdim = length(dim(x)),
  na.pad = TRUE,
  differences = 1L,
  ...
)

growth(x, lag = 1L, whichdim = length(dim(x)), logcompounding = FALSE)
}
\arguments{
\item{x}{an md3 object, or anything that may be converted to an md3}

\item{k, lag}{Integer or character. If integer, the number of backward lags used (or if negative the number of forward lags). Can also be a character time differnce corresponding to \code{timo}, such as \code{2M} for two months or \code{1W} for one week}

\item{whichdim}{which dimension of x to lag over (referred to as integer number of dimension name). By default the last dimension (typically \code{TIME}) but could be any other.}

\item{constrange}{logical. If TRUE, then the time range is expande to reflect the 'new'  periods. If FALSE (default), only time periods originally present in x will be preserved.}

\item{na.pad}{logical. IF TRUE (default) then the periods at the beginning o(ro end) of the original time range will be returned as NAs. See examples.}

\item{differences}{an integer for \code{diff} indicating the order of the difference.}

\item{logcompounding}{logical for \code{growth}, default FALSE. If TRUE, the result of \code{growth} is akin to log differences.
\dots not used}
}
\value{
an md3 object
}
\description{
Lags,  differences, and growth rates of MD3 objects akin to the lag function operating on other R objects
}
\examples{
xx=euhpq[TOTAL.I15_Q.AT:CY.y2005:y2008]
xx

lag(xx,k=4) #lagging by four quarters
lag(xx,k=4, constrange=FALSE) #lagging by 4 Q, and creating new time periods
lag(xx,k=4, na.pad=FALSE) #lagging by 4 Q, and not keeping old time periods

lag(xx,k='6M') #lagging by 6 months (using another frequency)
lag(xx,k=-4) #reverse lagging

growth(xx,lag = 4) #year-on-year growth rates (careful: these are not in percent)
diff(xx,lag = 4) #year-on-year differences

lag(xx,whichdim = 'geo') # lagging along countries
lag(xx,whichdim = 'geo', constrange=FALSE) # lagging along countries with constrange off

}
\seealso{
\code{\link{zapply}}, \code{\link[zoo]{rollapply}},  \code{\link{rollmean.md3}}
}
