% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3timo64.R
\docType{data}
\name{MD3-settings}
\alias{MD3-settings}
\alias{.cttim}
\title{Settings and options for timo time object}
\format{
An object of class \code{list} of length 7.
}
\usage{
.cttim
}
\description{
Settings and options for timo time object
}
\details{
hehe
}
\section{Indexing in md3}{
NA
}

\section{observation attributes}{
NA
}

\keyword{datasets}
