% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MD3.R
\name{MD3}
\alias{MD3}
\title{MD3: A package for handling multidimensional macro data with metadata}
\description{
MD3: A package for handling multidimensional macro data with metadata
}
\section{MD3 functionality}{

Check help(indexMD3) to see how to deal with MD3 objects
Use function \code{Nomics} and function \code{helpNomics} to load data
}

\section{Further work}{


This package is still under development; there are certainly some bugs lurking around
Build date: 2025-01-17


Conversion to and from \code{data.table}, tibble, and \code{pdata.frame} is to be added
}

\section{Author}{

Stefan Zeugner ECFIN.B1
}

