% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3utilities.R
\name{merge.md3}
\alias{merge.md3}
\title{Concatenates two or more md3 objects}
\usage{
\method{merge}{md3}(
  x,
  y,
  ...,
  along = NULL,
  newcodes = character(0),
  overwrite = NULL,
  ignoreDN = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{x}{an MD3 object with n dimensions}

\item{y}{an MD3 object with n or n-1 dimensions}

\item{\dots}{further MD3 objects with n or n-1 dimensions}

\item{along}{name of an existing or new dimension along which to combine x and y. If unspecified, then it tries to guess along whcih deimnsion to paste, depending on the value of \code{overwrite}}

\item{newcodes}{name of element names for the thing to be added (see \code{\link{dimcodes}}).}

\item{overwrite}{relevant if x and y have the same number of dimensions. if TRUE, then key combinations from y are used to overwrite key combinations from x. If FALSE, then the function tries to resolve conflicts by adding a new dimension to x}

\item{ignoreDN}{if FALSE (not default) then x and y are merged according to their dimension names, no matter the order of such dimensions (i.e. GEO is equal to GEO, but not equal to COUNTRY). If \code{FALSE} and  \code{x} and  \code{y} have the same number of dimensions, then  \code{merge.md3} matches instead the 1st to the 1st , the 2nd to the 2nd dimension, etc.}

\item{verbose}{if FALSE, suppress  success messages from the merging}
}
\value{
an md3 object with n or n+1 dimensions
}
\description{
Concatenates two or more md3 objects
}
\examples{

a1=euhpq['TOTAL','I15_Q',1:3,]

a2=euhpq[DW_NEW.I15_Q.BG.y:y2010]*2
myres=merge(a1,a2) #assumes this is a new country, adds a new country X1
print(myres)

myres=merge(a1,a2, newcodes='AB') #same but with a code
myres

merge(a1,a1['BE',,drop=FALSE]*2, overwrite=TRUE) #overwite Belgium with Belgium*2. Compare with a1
a1['BE',] = a1['BE',] *2
a1

merge(a1,a2, overwrite=FALSE) #is the same as
c(a1,a2)

}
\seealso{
\code{\link{drop.md3}}, \code{\link{aggregate.md3}}
}
