% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3nomics.R
\name{Nomics}
\alias{Nomics}
\alias{helpNomics}
\title{Load data from db nomics world}
\usage{
Nomics(
  query,
  as = c("md3", "2d", "1d", "array", "zooreg", "pdata.frame"),
  drop = NA,
  ccode = getOption("defaultcountrycode"),
  flags = FALSE,
  silent = FALSE,
  ...
)

helpNomics(query, pattern = "", dim = NULL, verbose = TRUE)
}
\arguments{
\item{query}{a character query string with a RestFul SDMX query (such as ECB/EXR/A.GBP+CHF.EUR.SP00.A). See Details}

\item{as}{how to as the result: md0: as md0 object with full metadata, 2d: as a data.frame with periods as column names, 1d: as fully stacked data.frame, \code{array} as multi-dim array,  \code{zooreg} as \code{zoo}-type time series,  \code{pdata.frame} as panel data frame from the \code{plm} package, with time and countries along rows, and all other indicators along columns}

\item{drop}{if TRUE, drop any singleton dimensions (see also \code{drop.md0})}

\item{ccode}{If this is not \code{NULL}, the the function will attempt to indetfy any coountrycode and try to convert them to the desired format (e.g. \code{iso3c} or \code{ec})}

\item{flags}{omit metadata, currently not operational}

\item{silent}{do not post messages about loading progress}

\item{...}{etc.}

\item{pattern}{pattern to match against the dataflow id or description. This uses in-built REgex expresions, but remains case-insensitive. If a pattern is not provided, all dataflows (or codes with labels) are returned .}

\item{dim}{dimension name or number (integer) denoting the dimension for which all codes and labels should be listed}
}
\value{
an md0, data.frame, or array, as specfied by argument 'as'.
Note that if used currectly, helpNomics invisibly returns either an md0 object, or lists with definitions
}
\description{
Load data from db nomics world
}
\details{
A restful SDMX query combines Provider, Dataflow and Dimension selectors
Take the example 'ECB/EXR/A.GBP+CHF.EUR.SP00.A'
\itemize{
 \item 'ECB' designates the provider ECB (see \code{helpNomics()} for a list of providers)
 \item 'EXR' designates the ECB's dataflow 'EXR' about exchange rates (see \code{helpNomics('ECB')} for a list of ECB dataflows)
 \item A.GBP+CHF.EUR.SP00.A is the dimension selector: See \code{helpNomics('ECB/EXR')} for what dimensions exist in ECB/EXR)
 }
 Here,
 \itemize{
 \item 'A'  denotes annual frequency from the FREQ dimension
 \item 'GBP+CHF' denote British pound and Swiss frank from the currency dimension
 \item 'EUR' denotes the currency denominator
 \item 'SP00' denots spot rates
 \item 'A' denotes average over periods (here annual averages)
 }
Omit dimension codes to download all elements of a dimension, i.e.
ECB/EXR/A..EUR.SP00.A to download exchange rates of the euro vis-a-vis all currencies

Note that a query is normally taken as Provider/Dataflow/Selectors, but Provider.Dataflow.Selectors is also tolerated.

Note also that db.nomics queries are case-sensitive
}
\section{Finding query codes}{


To find query codes, the easiest is to go to \url{http://db.nomics.world} and search for codes there
Alternatively, you can use \code{\link{helpNomics}}

\itemize{
 \item \code{helpNomics()} returns a vector with available providers
 \item \code{helpNomics("PROVIDER")} returns a vector with the dataflows available for the provider
 \item \code{helpNomics("PROVIDER", pattern="PATTERN")} looks for a dataflow whose name or code matches the string pattern
 \item \code{helpNomics("PROVIDER/DATAFLOW")} prints the structure of the dataflow dimensions, as well as an example query, and invisible returns the structure of code names within the dataflow
 \item \code{helpNomics("PROVIDER/DATAFLOW/SELECTORS")} invisibly returns the query result. If not successful, it provides clues whwere the query might be wrong
 \item \code{helpNomics("PROVIDER/DATAFLOW", dim="NumberOrName")} prints the code names and labels of the selected dimension, and invidibly returns a data.frame with those codes and labels
 \item \code{helpNomics("PROVIDER/DATAFLOW", dim="NumberOrName", pattern="PATTERN")} searches the codes and labels of that dimension for a certain pattern
}
}

\examples{
w1=Nomics('ECB/EXR/A..EUR.SP00.E')
print(w1[.y2018])

w2=Nomics('ECB/EXR/A..EUR.SP00.')
w2[ARS..y2012:y] #show Argentina since 2012

w3=Nomics('ECB/EXR/A.USD+JPY+CNY.EUR.SP00.E',as="2d")
w3

w4=Nomics('ECB/EXR/A.PLN.EUR.SP00.E',as="1d")
w4

helpNomics('ECB/EXR')
helpNomics('ECB/EXR/A.PLN.EUR.XXX.E')

#show available codes for fifth dimension
xx=helpNomics('ECB/EXR', dim=5)

#looking for the 'output gap' indicator in OECD's Economic Outlook dataflow
helpNomics('OECD/EO', dim='VARIABLE', pattern='gap')

}
