% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3plot.R
\name{plot.md3}
\alias{plot.md3}
\alias{lines.md3}
\title{Plot md3 time series}
\usage{
\method{plot}{md3}(
  x,
  y = NULL,
  plot.type = c("single", "multiple"),
  xlab = NULL,
  ylab = NULL,
  ...,
  plotflags = FALSE
)

\method{lines}{md3}(x, y = NULL, ...)
}
\arguments{
\item{x, y}{an md3 object}

\item{plot.type}{either 'single' or 'multiple'. Also wiorks by partial matching ('s' or 'm')}

\item{...}{arguments passed on to \code{\link[stats]{plot.ts}}}

\item{plotflags}{boolean. If TRUE then the charts indicats any flags (obsrvation status) on top of the data}
}
\description{
Plotting method for objects inheriting from class \code{md3}. Converts md3 to time series of class \code{ts} and applies \code{\link[stats]{plot.ts}}
}
\examples{
data(euca)
plot(euca[PC_GDP.SE.:2010])
plot(euca[PC_GDP.AT+SE.])
plot(euca[PC_GDP.AT+SE.],plot.type="single",col=1:2)
plot(euca[PC_GDP.LT.],euca[PC_GDP.LV.], xlab="Lithuania",ylab="Latvia")
lines(euca[PC_GDP.FR.],lwd=2,col='blue')

#barplot contribution of major countries to euro area current account
barplot(euca[MIO_NAC.DE+FR+IT+ES.])
}
\seealso{
\code{\link[stats]{plot.ts}} for the underlying method, \code{\link{barplot.md3}}
}
