% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3timo64.R
\name{seq.timo}
\alias{seq.timo}
\title{Sequencing}
\usage{
\method{seq}{timo}(from, to, by = NULL, frq = NULL, length.out = NULL, ...)
}
\arguments{
\item{from}{a starting \code{\link{timo}}  period. Required.}

\item{to}{end period (\code{\link{timo}} or any character, Date etc. that can be converted to a timo.  Optional.}

\item{by}{increment of the sequence. Integer. Default = 1}

\item{frq}{instead of by, use here the frequency code (like 'Q' or 'D') to define increments}

\item{length.out}{integer, optional. Desired length of the sequence.}
}
\value{
a timo vector
}
\description{
Sequencing
}
\details{
The following denotes the existing frequency codes and their notation
\itemize{
\item \code{A} annual, notation \code{'2011'}
\item \code{S} semi-annual, notation \code{'2011s2'}, second half of 2011
\item \code{Q} quarterly, notation \code{'2011q3'}, third quarter of 2011
\item \code{M} monthly, notation \code{'2011m10'}, October 2011
\item \code{W} weekly, notation \code{'2011w39'}, 39th week of 2011
\item \code{D} daily, notation \code{'2011-10-02'}, 2 October 2011
\item \code{N} minutely, notation \code{'2011-10-02t13:24'}, 2 October 2011, 13:24h (default GMT)
}
}
\examples{
seq(timo('2014m03'),'2015m06')
seq(timo('2014m03'),'2015')
seq(timo('2014m03'),'2015m06',by=3)
seq(timo('2014q1'),'2014q2',frq='W')
seq(timo('2014m03'),frq='B')
}
\seealso{
\code{\link{timo}}, \code{\link{frequency.timo}}, \code{\link{Sys.timo}}
}
