% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3timo64.R
\name{Sys.timo}
\alias{Sys.timo}
\title{Get current date and tiem as timo}
\usage{
Sys.timo(frq = NULL)
}
\arguments{
\item{frq}{optional paramter to be set to either NULL (default) or a frequency notation like "A", "S", "Q", "M","D", "B" or "N". Used for converting frequencies}
}
\description{
Get current date and tiem as timo
}
\details{
The timo object is a vector denoting mixed frequency periods like '2013', '2013m06', or '2013-06-24'.

The following denotes the existing frequency codes and their notation
\itemize{
\item \code{A} annual, notation \code{'2011'}
\item \code{S} semi-annual, notation \code{'2011s2'}, second half of 2011
\item \code{Q} quarterly, notation \code{'2011q3'}, third quarter of 2011
\item \code{M} monthly, notation \code{'2011m10'}, October 2011
\item \code{W} weekly, notation \code{'2011w39'}, 39th week of 2011
\item \code{B} business-day, notation \code{'2011-10-05'}, 5 October 2011
\item \code{D} daily, notation \code{'2011-10-02'}, 2 October 2011
\item \code{N} minutely, notation \code{'2011-10-02t13:24'}, 2 October 2011, 13:24h (default GMT)
}
}
\examples{
Sys.timo()
Sys.timo(frq='M')
Sys.timo(frq='W') + 0:5 #current week plus the next four ones
Sys.timo() -"1Q" # this minute one quarter ago

}
\seealso{
\code{\link{timo}}, \code{\link{frequency.timo}}, \code{\link{seq.timo}}
}
