% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3object.R
\name{t.md3}
\alias{t.md3}
\title{transpose a twodimensional md3 object}
\usage{
\method{t}{md3}(x)
}
\arguments{
\item{x}{an md3 object}
}
\value{
an md3
}
\description{
transpose a twodimensional md3 object
}
\examples{
#data(euhpq) #house prices for EU countries
euhpq[3,3,1:2,'2020:']
t(euhpq[3,3,1:2,'2020:'])

}
\seealso{
\code{aperm}
}
