% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3object.R
\name{time.md3}
\alias{time.md3}
\title{Extract the labels of the time dimensions}
\usage{
\method{time}{md3}(x, ...)
}
\arguments{
\item{x}{an md3 object}

\item{...}{not used, only there for compatitibilty reasons}
}
\value{
a timo object
}
\description{
Extract the labels of the time dimensions
}
\examples{

time(euhpq)
time(euhpq)+10
frequency(time(euhpq),'A') #convert frequency
euhpq[3,3,'AT',time(euhpq)-3] #3 year lag

}
\seealso{
\code{\link[stats]{time}}
}
