% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3utilities.R
\name{unflag}
\alias{unflag}
\title{Drop observation attributes}
\usage{
unflag(omd3, asDT = FALSE, attr2keep = "obs_value", ignoreNA = FALSE)
}
\arguments{
\item{asDT}{logical. Default \code{FALSE} returns an md3. \code{TRUE} returns a data.table}

\item{x}{an md3 object or stacked data.table}

\item{attr2keep\code{"value"}}{by default, but could also be "status" or "conf". If this is not 'value' or 'obs_value', asDT is assumed to be TRUE.}
}
\value{
an md3 object or dat.table, depending on \code{asDT}
}
\description{
By default, this removes obs attributes other than obs_value from an MD3 or associated data.table
}
\examples{

as.data.table(euhpq['TOTAL.I15_Q.BG.'])
as.data.table(unflag(euhpq['TOTAL.I15_Q.BG.']))
unflag(euhpq['TOTAL.I15_Q.BG.'], asDT=TRUE)
}
\seealso{
\code{\link{fill.md3}},  \code{\link{as.data.table}}
}
