% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3utilities.R
\name{aggregate.md3}
\alias{aggregate.md3}
\title{Aggregate an MD3 over time and/or multiple other dimensions}
\usage{
\method{aggregate}{md3}(
  x,
  frq_grp,
  along = "TIME",
  FUN = c(sum, mean, end, start),
  na.rm = TRUE,
  ...,
  complete.periods = !na.rm,
  drop = TRUE
)
}
\arguments{
\item{x}{an md3 object,}

\item{frq_grp}{what to aggregate to: A single character frequency code like \code{"A"},  \code{"Q"} or  \code{"B"} when aggregating along time, or a user-defined aggregate list (see details)}

\item{along}{character vector: the dimensions along which to aggregate. Default is \code{TIME}}

\item{FUN}{the function to use for aggregation. Default is \code{sum}, but \code{mean},  \code{end} and  \code{start} are also popular. Can be any function (passed along as a function or the string name of that function)}

\item{na.rm}{logical, default \code{TRUE}. Should missing values be removed?}

\item{\dots}{other arguments to FUN}

\item{complete.periods}{logical. Applies only in case of time aggregation and \code{na.rm=FALSE}. if e.g. \code{FALSE} and \code{FUN=end} then this takes the last value of the last available subperiod in case x ends within a period  (see examples).}

\item{drop}{Logical, default \code{TRUE}. See \code{\link{drop.md3}}}
}
\value{
an md3 object. Note that any flags or observation metadata from the original MD3 will be dropped and not contained in the resulting md3.
}
\description{
Can sum, average, etc. to lower frequencies, or along entire dimensions, or along user-defined aggregates (such as country groups)
}
\details{
For details on FUN see

Parameter \code{frq_grp} mostly is a single frequency code. See also \code{\link{frequency.timo}}.
The following denotes the existing frequency codes and their notation
\itemize{
\item \code{A} annual, notation \code{'2011'}
\item \code{S} semi-annual, notation \code{'2011s2'}, second half of 2011
\item \code{Q} quarterly, notation \code{'2011q3'}, third quarter of 2011
\item \code{M} monthly, notation \code{'2011m10'}, October 2011
\item \code{W} weekly, notation \code{'2011w39'}, 39th week of 2011
\item \code{D} daily, notation \code{'2011-10-02'}, 2 October 2011
\item \code{N} minutely, notation \code{'2011-10-02t13:24'}, 2 October 2011, 13:24h (default GMT)
}
}
\examples{
#data("oecdgdp_aq")

#Aggregating along the time dimension

yy=oecdgdp_aq['Q','AU:BE',,'2011q2:2015q2']
yy[AU..y2011q2+y2011q3]=NA #make some artificial NAs for demo purposes
print(yy[,'GDP_USD',])
aggregate(yy,'A')[,'GDP_USD',] # sum quarterly GDP figures to annual ones
aggregate(yy,'A',FUN=sum, na.rm=FALSE)[,'GDP_USD',] # sum quarterly GDP figures to annual ones.
#Note the NAs for 2011 and 2015 because 2011q1 and 2015q3:2015q4 are missing completely

#Alternatively sum those years where not all periods are available by setting complete.periods=FALSE
aggregate(yy,'A',FUN=sum, na.rm=FALSE, complete.periods=FALSE)[,'GDP_USD',]


aggregate(yy,'A',FUN=end)[,'GDP_USD',] #use the last period only
aggregate(yy,'A',FUN=end, na.rm=FALSE)[,'GDP_USD',] #use the last period only, but be careful about NAs

#Aggregating along other dimensions

aggregate(yy,along=c('LOCATION','TIME'),FUN=mean) # average GDP of countries
aggregate(yy,along=c('LOCATION'),FUN=sum) #GDP time series, sum of all three countries in yy

aggregate(oecdgdp_aq['Q..GDP_USD.y2015q1:'],list(LOCATION=list(G7=c('CA','DE','ES','FR','IT','JP','UK','US'),Baltics=c('EE','LV', 'LT'))))
aggregate(oecdgdp_aq['Q..GDP_USD.y2015q1:'],list(LOCATION=list(G7=c('CA','DE','ES','FR','IT','JP','UK','US'),Baltics=c('EE','LV', 'LT'))), along = 'LOCATION') # along is superfluous here
aggregate(oecdgdp_aq['Q..GDP_USD.y2015q1:'],list(G7=c('CA','DE','ES','FR','IT','JP','UK','US')), along = 'LOCATION') # but along is not superfluous here,  because frq_grp does not contain enough info
aggregate(oecdgdp_aq['Q..GDP_USD.y2015q1:'],c('CA','DE','ES','FR','IT','JP','UK','US'), along = 'LOCATION') # but along is not superfluous here,  because frq_grp does not contain enough info
aggregate(oecdgdp_aq['Q..GDP_USD.y2015q1:'],list(LOCATION=list(G7=c('CA', 'DE','ES','FR','IT','JP','UK', 'US'),Baltics=c('EE','LV', 'LT')),  TIME=character()), FUN=mean) #mean GDP across Time

#data(eupop)
aggregate(eupop['...y2019:y'],list(sex=list(all=c('F','M')),age=list(nwa=c('Y_LT15','Y_GE65'))))
#non-working age people in EU countries

}
\seealso{
\code{\link{fill.md3}},  \code{\link{disaggregate}}, \link{indexMD3}, \code{\link{imputena}}
}
