% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3plot.R
\name{barplot.md3}
\alias{barplot.md3}
\title{Plot md3 time series}
\usage{
\method{barplot}{md3}(
  height,
  ...,
  main = NULL,
  col = NULL,
  legend = TRUE,
  legend.text = NULL,
  legend.cex = 0.7,
  ylim = NULL
)
}
\arguments{
\item{height}{an md3 object, zoo, or matrix}

\item{...}{arguments passed on to \code{\link[graphics]{barplot}}}

\item{main}{title}

\item{col}{if left empty, then a grey colour palette}

\item{legend}{if TRUE, plots a legend for bars below the plot area}

\item{legend.text}{custom text for legend}

\item{legend.cex}{cex paramter for legend}
}
\value{
a vector with mid-points for bars. USe for additional plotting
}
\description{
Plotting method for objects inheriting from class \code{md3}. Converts md3 to time series of class \code{ts} and applies \code{\link[graphics]{barplot}}
}
\examples{
data(euca)
barplot(euca[PC_GDP.SE.:2010])

#barplot contribution of major countries to euro area current account
temp=barplot(euca[MIO_NAC.DE+FR+IT+ES+NL.]/1000, ylim=c(-100,500), ylab='bn EUR')
eaca=apply(euca['MIO_NAC',cgrp(EA),],2,sum,na.rm=TRUE)/1000
lines(temp,eaca,col='red',lwd=2) # adding a line with EA CA

}
\seealso{
\code{\link[stats]{plot.ts}} for the underlying method, \code{\link[stats]{sd}}
}
