% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3utilities.R
\name{disaggregate}
\alias{disaggregate}
\title{Disaggregate an MD3 over time and/or  other dimensions}
\usage{
disaggregate(x, frq_grp, along = "TIME", FUN = c(sum, mean, end, start), ...)
}
\arguments{
\item{x}{an md3 object,}

\item{frq_grp}{what to disaggregate to: A single character frequency code like \code{"M"},  \code{"W"} or  \code{"B"} when aggregating along time, or a user-defined aggregate list (see \link{frequency.timo})}

\item{along}{character vector: the dimensions along which to aggregate. Default is \code{TIME} - Note that all otehr dimensions are not possible for the moment}

\item{FUN}{the function to use for aggregation. Default is \code{sum}, but \code{mean},  \code{end} and  \code{start} are also possible (passed along as a function or the string name of that function)}

\item{\dots}{other arguments to FUN}
}
\value{
an md3 object. Note that any flags or observation metadata from the original MD3 will be dropped and not contained in the resulting md3.
}
\description{
Can decompose sums, averages, etc. to higher frequencies, or to constituent parts (e.g. countries in country groups)
}
\examples{
#data("oecdgdp_aq")

#Take some quarterly data: e.g. house price index for Austria and Belgium
x0=euhpq[TOTAL.I15_Q.AT+BE.y2020:y]
x0

#now disaggregate to monthly data
x1=disaggregate(x0,'M', FUN=mean)
x1
plot(x1)


}
\seealso{
\code{\link{fill.md3}}, \link{indexMD3}, \code{\link{imputena}}, \code{\link{aggregate.md3}}
}
