% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3object.R
\name{flags}
\alias{flags}
\alias{flagset}
\title{get or set observation attributes}
\usage{
flags(x, ...)

flagset(x, value, envir = parent.frame())
}
\arguments{
\item{x}{an md3 object}

\item{value}{the value to be set in case of \code{flagset}. Usually, a one-letter string.}

\item{envir}{environment for a call to \code{flagset}. Only there to handle exceptional cases}

\item{\ldots}{identifiers (see \link{indexMD3})}
}
\description{
puts flags like B for break, or confidentiality status
}
\section{Functions}{
\itemize{
\item \code{flagset()}: Setting \code{flags} for MD3 objects

}}
\examples{
testmd=euhpq['TOTAL.I15_Q.AT:BG.2021:']
flags(testmd) # shows pre built flags
testmd[.obs='status'] # shows pre built flags the same way

#this shows flags for sub-selections:
testmd['AT+BG.',.obs='status']
testmd['AT+BG.',.obs='flag'] #'flag' is another word for 'status'
testmd[c('AT','BG'),,.obs='flag']
flags(testmd['AT+BG.'])

#this way you can set flags
testmd['.2021', .obs='flag'] <- 'Q' #fail-safe way

flagset(testmd['.2021'], 'X') #convenient, but might face issues with identifying the correct encapsulating environment
testmd['AT.2021q2', .obs='flag'] <-'z'

#see the result:
flags(testmd)

testdd=as.data.table(testmd) #convert to data.table
testdd[ TIME=='2022q4', obs_status:='m']
testmd2=as.md3(testdd) #converting back to md3
flags(testmd2)


# Note that in assigning on from one MD3 to another such as euhpq["TOTAL.RCH_A.SK."] <-euhpq["TOTAL.RCH_A.FR.",]
# by default only the data value will be assigned.

#Selecting a specific .obs assigns that specific .obs only:
\dontrun{euhpq["TOTAL.RCH_A.SK.",.obs="flag"] <-euhpq["TOTAL.RCH_A.FR.",.obs="flag"]}

#show the result:
\dontrun{euhpq["TOTAL.RCH_A.SK.",.obs="flag"]}

#But .obs can also be a vector to assign values and multiple attributes at once, such as in

\dontrun{euhpq["TOTAL.RCH_A.SK.",.obs=c("value","flag")] <-euhpq["TOTAL.RCH_A.FR."]}

#or by setting .obs="all"

\dontrun{euhpq["TOTAL.RCH_A.SK.",.obs="all"] <-euhpq["TOTAL.RCH_A.HU."]}

#Finally, note that data attributes can only be carried if the value of that observation is not NA or NaN. The value of the observation can be Inf, however.




}
\seealso{
\code{\link{indexMD3}}
}
