% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3object.R
\name{flags}
\alias{flags}
\title{get or set observation attributes}
\usage{
flags(x, ...)
}
\arguments{
\item{x}{an md3 object}

\item{\ldots}{identifiers (see \link{indexMD3})}
}
\description{
puts flags like B for break, or confidentiality status
}
\examples{
testmd=euhpq['TOTAL.I15_Q.AT:BG.']
flags(testmd)
flags(testmd['.2015']) <- 'T'

as.data.table(testmd)

}
\seealso{
\code{\link{indexMD3}}
}
