% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3object.R
\name{indexMD3}
\alias{indexMD3}
\alias{[.md3}
\title{Extract or replace parts of an md3 object}
\usage{
\method{[}{md3}(
  x,
  ...,
  drop = TRUE,
  as = c("md3", "array", "numeric", "data.table"),
  .obs = "value",
  DT = FALSE
)
}
\arguments{
\item{x}{an md3 object}

\item{...}{indexes, see 'Details' below}

\item{drop}{whether to drop singleton dimensions (see \code{\link{drop.md3}})}

\item{as}{how to return the result, see also \code{\link{as.data.table.md3}}}

\item{.obs}{used when extracting or replacing not the observation values, but observation attributes such as flags (see below)}

\item{onlyna}{only update those elements that are NA}

\item{justval}{only update those elements that are not NA}

\item{usenames}{when replacing parts of x, update x by using the dimension codes, rather than the order, of elements in value}
}
\value{
an md3, array, numeric, zoo, or data.table, data.frame depending on paramter \code{as}
}
\description{
Extract or replace parts of an md3 object
}
\details{
and md3 object can be indexed like an array (\code{euhpq['TOTAL',2,7:9,'2021q1']}),
or in an SDMX API-like notation (\code{euhpq['TOTAL.I15_Q.BE+FR.2021q1:']} resp (\code{euhpq['TOTAL..SI.2021q1:']}),
or a mix thereof (\code{euhpq[3,2,'BE+FR','2021q1:']})
}
\section{Internal structure}{

an md3 actually is a data.table that contains one column for each dimension,
and at least one column for observations (there can be additional columns for flags and confidentiality labels).
You can see that  by running .e.g \code{as.data.table(euhpq)}
}

\seealso{
\code{\link{dimcodes}}
}
